Change to ExchangeService and tests

+++ src/test/java/org/prebid/server/handler/openrtb2/AmpHandlerTest.java
@@ -440,7 +440,7 @@ public class AmpHandlerTest extends VertxTest {
         given(exchangeService.holdAuction(any()))
                 .willReturn(givenBidResponseWithExt(mapper.valueToTree(
                         ExtBidResponse.of(ExtResponseDebug.of(null, auctionContext.getBidRequest()), null, null, null,
-                                null, (long )1000))));
+                                null, 1000L))));
 
         // when
         ampHandler.handle(routingContext);
